pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "quantflow"
version = "0.1.0"
description = "A-share quantitative trading system (live + backtest)"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "pandas>=2.0",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "tqdm>=4.65",
]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "ruff>=0.1",
    "black>=23.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-ra", "-q", "--strict-markers"]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.black]
line-length = 120
target-version = ["py313"]
